Struct crosvm::argument::Argument

pub struct Argument {
    pub value: Option<&'static str>,
    pub value_mode: ArgumentValueMode,
    pub short: Option<char>,
    pub long: &'static str,
    pub help: &'static str,
}

Information about an argument expected from the command line.

Examples

To indicate a flag style argument:

Argument::short_flag('f', "flag", "enable awesome mode");

To indicate a parameter style argument that expects a value:

// "VALUE" and "NETMASK" are placeholder values displayed in the help message for these
// arguments.
Argument::short_value('v', "val", "VALUE", "how much do you value this usage information");
Argument::value("netmask", "NETMASK", "hides your netface");

To indicate an argument with no short version:

Argument::flag("verbose", "this option is hard to type quickly");

To indicate a positional argument:

Argument::positional("VALUES", "these are positional arguments");

Fields
value: Option<&'static str>

The name of the value to display in the usage information.

value_mode: ArgumentValueMode

Specifies how values should be handled for this this argument.

short: Option<char>

Optional single character shortened argument name.

long: &'static str

The long name of this argument.

help: &'static str

Helpfuly usage information for this argument to display to the user.

Implementations

impl Argument

pub fn positional(value: &'static str, help: &'static str) -> Argument

pub fn value(
long: &'static str,
value: &'static str,
help: &'static str
) -> Argument

pub fn short_value(
short: char,
long: &'static str,
value: &'static str,
help: &'static str
) -> Argument

pub fn flag(long: &'static str, help: &'static str) -> Argument

pub fn short_flag(
short: char,
long: &'static str,
help: &'static str
) -> Argument

pub fn flag_or_value(
long: &'static str,
value: &'static str,
help: &'static str
) -> Argument