Module crosvm::crosvm::plugin

Modules

process

vcpu

Structs

VcpuPipe

Enums

CommError

An error that occurs when communicating with the plugin process.

PluginObject

Each PluginObject represents one object that was instantiated by the guest using the Create request.

Token

Constants

CROSVM_GPU_SERVER_FD_ENV

MAX_DATAGRAM_SIZE

MAX_VCPU_DATAGRAM_SIZE

Functions

create_plugin_jail

io_to_sys_err

mmap_to_sys_err

new_pipe_pair

new_seqpacket_pair

proto_to_sys_err

run_config

Run a VM with a plugin process specified by cfg.

run_vcpus