Crate crosvm

Runs a virtual machine

Modules

crosvm

The root level module that includes the config and aggregate of the submodules for running said configs.

panic_hook

sys

Enums

CommandStatus

Functions

balloon_stats

balloon_vms

create_composite

create_qcow2

crosvm_main

disk_cmd

inject_gpe

main

make_rt

modify_battery

modify_usb

modify_vfio

pkg_version

powerbtn_vms

resume_vms

run_vm

sleepbtn_vms

start_device

stop_vms

suspend_vms

usb_attach

usb_detach

usb_list