Crate crosvm

The root level module that includes the config and aggregate of the submodules for running said configs.

Modules

argument

Handles argument parsing.

gdb

platform

plugin

Structs

BindMount

A bind mount for directories in the plugin process.

Config

Aggregate of all configurable options for a running VM.

DirectIoOption

Direct IO forwarding options

DiskOption

GidMap

A mapping of linux group IDs for the plugin process.

SharedDir

TouchDeviceOption

VfioCommand

VFIO device structure for creating a new instance based on command line options.

VhostUserFsOption

VhostUserOption

VhostUserWlOption

Enums

Executable

Indicates the location and kind of executable kernel for a VM.

SharedDirKind

VfioType

Vfio device type, recognized based on command line option.

VhostVsockDeviceParameter

Constants

DEFAULT_TOUCH_DEVICE_HEIGHT

DEFAULT_TOUCH_DEVICE_WIDTH

DISK_ID_LEN

Maximum length of a DiskOption identifier.